update staking client structs to adhere to the model convention
